add readonly and move app to a folder
