+++ src/features/app/App.tsx
@@ -1,13 +1,13 @@
 import React from "react";
 import { useSelector } from "react-redux";
-import { Headline } from "./features/headline";
-import { Chart } from "./features/chart";
-import { LatestNews } from "./features/latest-news";
-import { KeyStats } from "./features/key-stats";
-import { Company } from "./features/company";
-import { Footer } from "./features/footer";
+import { Headline } from "../headline";
+import { Chart } from "../chart";
+import { LatestNews } from "../latest-news";
+import { KeyStats } from "../key-stats";
+import { Company } from "../company";
+import { Footer } from "../footer";
 import "./App.css";
-import { AppState } from "./store";
+import { AppState } from "../../store";
 
 const App = () => {
   const selectedStock = useSelector(
