+++ src/index.tsx
@@ -2,7 +2,7 @@ import React from "react";
 import ReactDOM from "react-dom";
 import "./styles/Global.css";
 import "./styles/Variables.css";
-import App from "./App";
+import App from "./features/app/App";
 import * as serviceWorker from "./serviceWorker";
 
 import { bootstrap } from "./store/actions";
