+++ src/store/constants.ts
@@ -1,3 +1,2 @@
 export const BOOTSTRAP = "BOOTSTRAP";
 export const RESET = "RESET";
-export const UPDATE_SELECTED_STOCK = "UPDATE_SELECTED_STOCK";
